Enabling PTP clock support in Linux Kernel configs
The IEEE 1588 standard defines a method to precisely
synchronize distributed clocks over Ethernet networks. The
standard defines a Precision Time Protocol (PTP), which can
be used to achieve synchronization within a few dozen
microseconds. In addition, with the help of special hardware
time stamping units, it can be possible to achieve
synchronization to within a few hundred nanoseconds.

Story: 2002935
Task: 22922

Change-Id: Ibb4e9b8f61198c88d3aaec02f441574b580afdd7
Signed-off-by: Alex Kozyrev <alex.kozyrev@windriver.com>