CentOS 8: Fix spec file for qat17
(1) Add one BuildRequires: elfutils-libelf-devel,
because the elfutils-libelf-devel package contains the
libraries to create applications for handling compiled objects.

(2) Changed the signatures of the modules method,
because sign-file on centos 8 is not a perl script,
but an executable command.
The file extension of the private key was changed from priv to pem.

Change-Id: Ie8a7f851e97a5a7014983a0528c466dfe89b790a
Story: 2006728
Task: 37663
Depends-On: https://review.opendev.org/#/c/695134/
Depends-On: https://review.opendev.org/#/c/696481/
Depends-On: https://review.opendev.org/#/c/696050/
Signed-off-by: Zhang Zhiguo <zhangzhg@neusoft.com>