CentOS 8: Upgrade intel-e1000e to version 3.6.0
Fix spec file for centos 8:

(1) Changed the signatures of the modules method,
because sign-file on centos 8 is not a perl script,
but an executable command.
The file extension of the private key was changed from priv to pem.

(2)Upgrade tarball version from 3.4.2.1 to 3.6.0,
because the old version is missing support to build against CentOS 8.

Change-Id: I592ca3b2a6287bc3b10532228b34d0095fa536be
Story: 2006728
Task: 37662
Depends-On: https://review.opendev.org/#/c/695134/
Depends-On: https://review.opendev.org/#/c/696481/
Depends-On: https://review.opendev.org/#/c/696050/
Signed-off-by: Zhang Zhiguo <zhangzhg@neusoft.com>