Build mpt2sas and mpt3sas drivers as modules
History:
Back in the day, we didn't have an initramfs
to allow us to load disk drivers as modules.  All
disk drivers had to be built-in.  In CentOS 7.3,
the mpt2sas and mpt3sas driver code was reorganized
to allow for a common code base.  But along with that,
those drivers would only now build as modules.  We
created a patch which involved taking a snapshot of
mpt driver code, and massaged it all into building
as built-in drivers.

Problem:
That old code snapshot along with the fact
that those two drivers initialize without their
associated hardware being present (they are built-in),
seems to cause interference with some other LSI raid
controllers, namely Harpoon in AVAGO MR9460-8i via a
Huawei enclosure.

Solution:
Simply revert to building those two mptsas drivers as
modules, the way CentOS intended.  They will reside
on initramfs and be loaded automatically if the
appropriate hardware is present. With these drivers now
out of the way, the problematic raid controller works
fine, driven by the megaraid_sas driver.

Change-Id: I054c2396df4e659c324e70bffcf3940ad93c9354
Closes-Bug: 1866293
Signed-off-by: Jim Somerville <Jim.Somerville@windriver.com>