Update iavf driver sign method with 4.18 kernel from CentOS 8
(cherry picked from commit b418537fae13e43140a1f5f2ac37b3e958461d7c)

Changed the signatures of the modules method,
because sign-file on centos 8 is not a perl script,
but an executable command.
The file extension of the private key was changed from priv to pem.

Change-Id: I4adb4c693b89d23122183a45b2daab12249e5982
Story: 2007308
Task: 38800
Depends-On: https://review.opendev.org/720325
Signed-off-by: Shuicheng Lin <shuicheng.lin@intel.com>
Signed-off-by: Dongqi Chen <chen.dq@neusoft.com>