Add kernel-rt to the centos_tarball-dl list
This adds the kernel-rt rpm to go along with the changes
in dl_tarball in the tools repo. The populate_downloads script
will correctly link the src.rpm into the stx/downloads directory.

The next step is to change the srpm_path in the 4.180-rt commit
to use the downloaded srpm:

repo:stx/downloads/kernel-rt-4.18.0-147.3.1.rt24.96.el8_1.src.rpm

This will cause the src.rpm to be fetched from the downloads directory
in order and then further patched by the build_srpm process to generate
the STX variation of kernel-rt src.rpm.

Depends-On: https://review.opendev.org/722017
Change-Id: Ic57761adcae878af8d44b622d81c65ab017c555a
Signed-off-by: Saul Wold <sgw@linux.intel.com>