Move tpmdd to built-in driver
The tpm driver is the same as what is now in the kernel,
aside from one patch.  So we stop building it as an
external module and instead use the one in the kernel.
Move that one patch to it to apply to the kernel code.

Test Methods:
Refer to https://wiki.openstack.org/wiki/StarlingX/Distro/tpm-verify
to verify the encryption and decryption function is available,
and the signature verification function is available.

Update the TIS_PATCH_VER to avoid an older
built version of a package being picked up.

Change-Id: I3bfcae4ab8c25778b29d51cc9dfb6cb1d81eef7d
Story: 2007308
Task: 39099
Signed-off-by: Dongqi Chen <chen.dq@neusoft.com>