Uprev kmods to fix layered build
The update that broke layered build was
https://review.opendev.org/#/c/727908

It upversioned the kernels, but failed to follow the guidance to
upversion all external kmods.

Build avoidance is in use for layered builds, seeing no source or
metadata changes for the kmod, it elected to reuse the previous build of
the kmod.  The previous kmod is only valid for the previous kernel due
to embedded paths that include the kernel version.  A kernel/kmod
mismatch that results in unusable kmods, no network drivers, and an
install failure on many pieces of hardware.

Build avoidance is not in use for the monolithic build.

Short term fix.  Upversion the kmods

Partial-Bug: 1878072
Change-Id: I6eb2cce6919ce61b07ea14e70081a6433e686630
Signed-off-by: Scott Little <scott.little@windriver.com>