Add support for igb_uio kernel module
Currently, the forward error correction (FEC) device on an
Intel N3000 FPGA card requires the use of the igb_uio driver
for use with DPDK applications.

This commit adds support for the igb_uio module by building
DPDK 20.05 and extracting the enabled driver into the kernel
modules.

In addition, DPDK's devbind script is installed in the system
at /usr/share/starlingx/scripts/dpdk-devbind.py to assist
in binding the driver to a device. It is not possible to
use the existing (older) version of this script shipped with
openvswitch as it does not support acceleration devices such
as the FEC.

Story: 2006740
Task: 39946
Change-Id: I838a6dacaaa79fcbce22884f42aa141435964b1d
Signed-off-by: Steven Webster <steven.webster@windriver.com>