Auto upversioning of kernel and kmods
when kernel or kmod is modified

The PKG_GITREVCOUNT component of TIS_PATCH_VER
will increment every time that a change is
delivered to $PKG_BASE for the package.
The actual value is the number of commits
since PKG_BASE_SRCREV, and an optional plus 1
if there are uncommitted changes.

Kmods now add OPT_DEP_LIST_FOR_BUILD_TYPE to list
the kernel's source tree as a non-built input
to each kmod.  If the kernel is modified,
the OTHER_GITREVCOUNT component of the
TIS_PATCH_VER will increment. The actual value is
the number of commits since BASE_SRCREV_FOR_PATH,
and an optioanal plus 1 if there are uncommitted
changes.

So a kernel only change will auto increment and
build not only the kernel, but all kmods.

A kmod change will only auto-increment the affected kmod.

If a package is rebased to a new upstream version,
then PKG_BASE_SRCREV, and BASE_SRCREV_FOR_PATH if present,
should be adjusted to drop TIS_PATCH_VER to 1.

Note: BASE_SRCREV_FOR_PATH for a kmod should NOT be modified
if the kernel is rebased to a newer upstream kernel, but
the kmod itself has not rebased at the same time.  Only adjust
BASE_SRCREV_FOR_PATH if the kmod itself has been rebased.
Why?  The kmod's TIS_PATCH_VER would decrement!  Yum
will prefer the older build of the kmod, built against
the old kernel.  The installed kmod and kernel will not match.

Story: 2007750
Task: 39937
Depends-On: https://review.opendev.org/734687
Change-Id: Ie8f840934b4ddfd4c09ea3dc738e4f4032683dfc
Signed-off-by: Scott Little <scott.little@windriver.com>