Upversion opae-intel-fpga-driver source to 2.0.1-8
Introduce the new source ahead of actually using it.  This
gives users a chance to download it in advance, to lessen
the likelihood of having loadbuilds get blocked on
waiting for downloads which can be a slow process.

The entry referring to the old source will be deleted
upon switching the build over to use the new.

Change-Id: I2b3a144f76e9d6aeeee185cf9ce13b2771c81cf5
Story: 2006740
Task: 40109
Signed-off-by: Jim Somerville <Jim.Somerville@windriver.com>