Drop weak-modules call from kmod RPM scripts
The kmod RPM scriptlets are calling weak-modules from the postinstall
and postuninstall scriptlets. Each call is taking approximately 40
seconds. On an AIO install, for example, this adds about 5 minutes to
the install. For a kernel software update, since both postinstall and
postuninstall scriptlets are called, this can add about 10 minutes to
the update installation.

As weak modules provide little benefit in a closed system, these calls
have been removed in this update.

Change-Id: I7ef577667bef1e75a0aa8542c76d401ecd5c896a
Closes-Bug: 1876500
Signed-off-by: Don Penney <don.penney@windriver.com>