qat: fix smp_processor_id preemption complaints
The module loading works but the code is evidently violating
the use of smp_processor_id() with respect to preemptibility.

It seems that smp_processor_id() is only used for a best-effort
load-balancing, refer to qat_crypto_get_instance_node(). It's not feasible
to disable preemption for the duration of the crypto requests. Therefore,
just silence the warning.

Reference materials:
https://github.com/torvalds/linux/commit/1b82feb6c5e1996513d0fb0bbb475417088b4954
https://www.kernel.org/doc/Documentation/preempt-locking.txt

Change-Id: I0f4d88d934aa29d30cde9a20212e758e15ad01ad
Closes-Bug: 1884262
Signed-off-by: zhao.shuai <zhaos@neusoft.com>