rdma-core/libibverbs: add initial versions for debian
Both debs for rdma-core and libibverbs are built by this
debian folder added. Their versions are aligned with mlnx-ofa_kernel
driver and all are from:
https://linux.mellanox.com/public/repo/mlnx_ofed/5.3-1.0.0.1/SRPMS
The source code package name for rdma-core and libibverbs is
rdma-core-52mlnx1.tgz.

Verification:
As we are only getting the Debian work bootstrapped there is quite a few
restrictions as far as what can be tested.
- build image ok with the module;
- qemu boot up ok with the module.

Test on hardware will be done when the integration of the debian
platform is done.

Story: 2009221
Task: 43337
Signed-off-by: Li Zhou <li.zhou@windriver.com>
Change-Id: I6593512985be9ff8e5c350a15cfd4017779d4093