intel-opae-fpga: add initial version for debian
Add kmod-opae-fpga-driver 2.0.1-10 of debian packaging.

The source code package is from:
opae-intel-fpga-driver-2.0.1-10.src.rpm in
https://github.com/OPAE/opae-sdk/releases/download/1.3.7-5/OPAE_1.3.7-5_el7.zip

The patches are carried over from the 5.10 kernel with stx CentOS by
20211108.
use-altera.h-from-driver-instead-of-from-kernel.patch is adjusted a
little for the header file path issue which causes building error.

Verification:
As we are only getting the Debian work bootstrapped there is quite a few
restrictions as far as what can be tested.
- source code/patch/config are aligned with stx centos 20211108;
- output packages' contents are aligned with stx centos 20211108;
- build image ok with the module;
- qemu boot up ok with the module.

Story: 2009221
Task: 43411
Signed-off-by: Li Zhou <li.zhou@windriver.com>
Change-Id: Ie7c2a28653e9a30f7c9aa1d161c9679f66ad6188