qat17: add initial version for debian
Add qat17 4.14.0-00031 driver of debian packaging.

The source code package is from:
https://downloadmirror.intel.com/30178/eng/QAT1.7.L.4.14.0-00031.tar.gz

The patches and files (qat/qat_service) in debian/extra folder are
carried over from the 5.10 kernel with stx CentOS by 20211108
(original files' path: kernel-modules/qat17/files).
Added a new patch to fix the linking error related to gcc used in stx
Debian.

Verification:
As we are only getting the Debian work bootstrapped there is quite a few
restrictions as far as what can be tested.
- source code/patch/config are aligned with stx centos 20211108;
- output packages' contents are aligned with stx centos 20211108;
- build image ok with the module;
- qemu boot up ok with the module.

Test on hardware will be done when the integration of the debian
platform is done.

Story: 2009221
Task: 43413
Signed-off-by: Li Zhou <li.zhou@windriver.com>
Change-Id: Id5a75341fd7a1b335b5525e3bfd288c7816e7ef3