mlnx-ofa_kernel: add initial version for debian
Add mlnx-ofa_kernel 5.3 of debian packaging.

The source code package is from:
https://linux.mellanox.com/public/repo/mlnx_ofed/5.3-1.0.0.1/SRPMS/mlnx-ofa_kernel-5.3.tgz

The patches are carried over from the 5.10 kernel with stx CentOS by
20211108.
0001-implicit-declaration-of-function-__is_constexpr.patch is adjusted
on the file path because of the patching failure.

Verification:
As we are only getting the Debian work bootstrapped there is quite a few
restrictions as far as what can be tested.
- source code/patch/config are aligned with stx centos 20211108;
- output packages' contents are aligned with stx centos 20211108;
- build image ok with the module;
- qemu boot up ok with the module.

Story: 2009221
Task: 43412
Signed-off-by: Li Zhou <li.zhou@windriver.com>
Change-Id: I21e8b49259afb95304e3830ff4f99cb5a6778cbb