mlnx-ofa_kernel: Remove mlx4* from modules-load.conf
This commit removes mlx4_ib, mlx4_core and mlx4_en modules from
mlx-ofa_kernel's modules-load.conf file.

This is necessary because the kernel modules are no longer built as part
of the package. They appear to have been deprecated, and the configure
script no longer supports the --with-mlx4-mod and --with-mlx4_en-mod
arguments.

Story: 2008921
Task: 43629

Signed-off-by: M. Vefa Bicakci <vefa.bicakci@windriver.com>
Signed-off-by: Jiping Ma <jiping.ma2@windriver.com>
Change-Id: I038413161c1c8e45ae638042ff5559c4c409832f