scsi: Make the disk detection order more consistent
Starting with the v5.x kernels asynchronous device scanning is enabled
by default to speed up the boot process. This results in device
detection order and thus naming potentially changing from one boot to
the next, affecting disk and USB storage device names.

We revert two related patches, returning the kernel device naming
convention of pre-v5.x kernels. Since we are slotting a 5.10 kernel into
a fairly old userspace, this is preferred approach versus adapting the
userspace to this new kernel behavior, at least at this time.

NOTE: BIOS and other factors outside the kernel's control may still
result in different device naming, users should be aware of this if
performing firmware updates. In the long term different approaches will
be required to determine which devices to use that don't rely on device
names, allowing the original 5.10 kernel behavior to work as designed.

Verification:
Ran test 120 times with an unpatched 5.10.74 kernel, 120 times with an
unpatched 5.10.74 kernel with "scsi_mod.scan=sync" and 220 times with a
patched 5.10.74 kernel with the two commits reverted. Here are the
results:

1. 5.10.74 unpatched without scsi_mod.scan=sync
   --14/120 instances of unexpected disk order.
2. 5.10.74 unpatched with scsi_mod.scan=sync
   --12/120 instances of unexpected disk order.
3. 5.10.74 patched without scsi_mod.scan=sync
   --220 instances of the same disk order.

Closes-bug: #1953433
Closes-Bug: #1950163

[Submitted on behalf of Vefa Bicakci.]
Signed-off-by: M. Vefa Bicakci <vefa.bicakci@windriver.com>
Signed-off-by: Jiping Ma <jiping.ma2@windriver.com>
Change-Id: I4f508fce99a30deb4ab6688682a1057205013679