Ice Driver: use new comms ddp firmware
This is a firmware package for Intel Ethernet 800 series
adapters, addressing dynamic device personalization for the
communications market.

Instead of using the default ice ddp firmware that
comes with the ice driver, we switch to use the
comms flavour of that ddp firmware.  It is in a
separate zip file which simply has to be unpacked
and then repackaged as part of the ice driver.

We looked at doing this as a separate ice_comms package
but it would have to be dracut maintained, not to
mention the possibility of filename collisions with the
default in the ice driver.  So we deemed it prudent to
just fold it into the ice driver package.

Of note is that the comms flavour is a superset of
the default, it includes everything in the default
plus more protocols.

Verification:
- boot from iso, check dmesg output to see that the
ice driver has loaded the ICE COMMS DDP FW and that it
reports the correct version
- verify that the ice hardware nic works correctly
through lab installation while using this new fw.  We
do not verify any of the newly added protocols.

Story: 2009823
Task: 44409
Depends-On: https://review.opendev.org/c/starlingx/tools/+/827571

Change-Id: I832b38cf5d579116f79b2a9be3332ad8ff2add3b
Signed-off-by: Jim Somerville <Jim.Somerville@windriver.com>