kernel: Add auxiliary bus support
This commit back-ports the auxiliary bus device driver from kernel
version 5.11 to the StarlingX kernels, along with the bug fixes to this
driver to date.

This is necessary, because Intel's out-of-tree ice and iavf drivers, and
Mellanox OFED v5.5's out-of-tree drivers rely on the functionality
provided by the auxiliary bus driver. However, because the v5.10 kernel
baseline used by StarlingX does not have this driver, the out-of-tree
drivers would build and install their own version of the driver, named
auxiliary.ko. This would cause kernel module symbol resolution errors at
run-time when attempting to load ice and mlx5 kernel modules
concurrently.

Hence, this commit cherry-picks eight patches to add the auxiliary bus
device driver to StarlingX's kernels, and adds a ninth patch to enable
CONFIG_AUXILIARY_BUS by default. The latter is necessary, because
StarlingX does not enable any in-tree kernel modules that 'select'
CONFIG_AUXILIARY_BUS, which causes the kernel's build system to
automatically disable this option. (None of the cherry-picked patches
had merge conflicts.)

It should be noted that all affected out-of-tree drivers' build systems
are robust enough to detect that the kernel provides an auxiliary.ko
driver, so the only out-of-tree driver modification involved changing
the ice driver's spec file so that the installation of auxiliary.ko is
removed.

Testing:
- An incremental monolithic build of StarlingX's master branch was
  successful. This build included currently-unreleased Mellanox OFED
  v5.5 work as well.

- The built ISO image was successfully installed into and bootstrapped
  in a VM with low-latency profile in All-in-One simplex configuration,
  and the ability to insert ice, iavf and mlx5 kernel modules
  concurrently was confirmed.

- The same test was repeated with a different VM, this time with the
  standard (i.e., non-low-latency) profile in All-in-One simplex mode.

- The same ISO image was installed onto a 'standard' StarlingX system
  with two controller and two compute nodes with network adapters
  managed by Mellanox's OFED drivers. The low-latency profile was used.
  The Mellanox network adapters are used for inter-node communication in
  this server set-up. The bootstrap procedure was successful, and the
  RDMA/Infiniband over Ethernet functionalities of the Mellanox adapters
  were successfully tested using the Linux RDMA community's perftest
  package.

- The same ISO image was installed onto an All-in-One simplex StarlingX
  system with an Intel E810 network adapter, managed by the ice driver,
  which is the system's management (OAM) interface. While the bootstrap
  of the system ultimately failed, this was due to an unrelated reason,
  and the network adapter's functionality was not found to be affected
  by this commit. iperf3 was used on virtual function interfaces
  (belonging to the same E810 network adapter) manually set up and
  attached to separate network namespaces, as an additional sanity test.

Story: 2009878
Task: 44612

Change-Id: I8acfb0539757a35be3fb099ea2a6f95e05557e4f
Signed-off-by: M. Vefa Bicakci <vefa.bicakci@windriver.com>